# api/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ETask API Documentation",
    "description": "基于 AST 与反射契约混合自适应生成的 OpenAPI 3.0 规范，支持导入 Apifox / Postman 进行全量接口联调。",
    "version": "1.0.0"
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/api/code-assist/change-set/apply": {
      "post": {
        "tags": [
          "脚本引擎/AI 助手"
        ],
        "summary": "应用 AI 项目变更",
        "description": "权限标识: apply_change_set",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyChangeSetReq"
              },
              "example": {
                "id": 1
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/api/code-assist/conversation/create": {
      "post": {
        "tags": [
          "脚本引擎/AI 助手"
        ],
        "summary": "创建 AI 会话",
        "description": "权限标识: add_conversation",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConversationReq"
              },
              "example": {
                "project_id": 1,
                "title": "示例名称"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/api/code-assist/conversation/detail": {
      "post": {
        "tags": [
          "脚本引擎/AI 助手"
        ],
        "summary": "AI 会话详情",
        "description": "权限标识: get_conversation",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConversationDetailReq"
              },
              "example": {
                "conversation_id": 1
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/api/code-assist/conversation/list": {
      "post": {
        "tags": [
          "脚本引擎/AI 助手"
        ],
        "summary": "AI 会话列表",
        "description": "权限标识: view",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListConversationsReq"
              },
              "example": {
                "project_id": 1
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/api/code-assist/message/stream": {
      "post": {
        "tags": [
          "脚本引擎/AI 助手"
        ],
        "summary": "发送 AI 消息",
        "description": "权限标识: chat",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatReq"
              },
              "example": {
                "content": "string",
                "context": "string",
                "conversation_id": 1,
                "profile_id": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
...
  },
...
  "components": {
    "schemas": {
...
      "ApplyChangeSetReq": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          }
        },
        "example": {
          "id": 1
        }
      },
...
      "ChatReq": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "example": "string"
          },
          "context": {
            "type": "string",
            "example": "string"
          },
          "conversation_id": {
            "type": "integer",
            "example": 1
          },
          "profile_id": {
            "type": "string",
            "example": "string"
          }
        },
        "example": {
          "content": "string",
          "context": "string",
          "conversation_id": 1,
          "profile_id": "string"
        }
      },
...
      "ConversationDetailReq": {
        "type": "object",
        "properties": {
          "conversation_id": {
            "type": "integer",
            "example": 1
          }
        },
        "example": {
          "conversation_id": 1
        }
      },
      "CreateConversationReq": {
        "type": "object",
        "properties": {
          "project_id": {
            "type": "integer",
            "example": 1
          },
          "title": {
            "type": "string",
            "example": "示例名称"
          }
        },
        "example": {
          "project_id": 1,
          "title": "示例名称"
        }
      },
...
      "ListConversationsReq": {
        "type": "object",
        "properties": {
          "project_id": {
            "type": "integer",
            "example": 1
          }
        },
        "example": {
          "project_id": 1
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "请输入登录后获取的 JWT Bearer Token 进行接口调用鉴权"
      }
    }
  }
}